Adds the basic paper structure using the AutRob format.
This is under the folder /report along with submodule CMakeLatex.

Latex build is cmake'd, so @Tabjones, if you use catkin-build you will
compile the article as well ;)... get well!

We can move it to another branch if it annoys the code building,
or add it to the target `make doc` of some package, like the gaussian
object modelling.
I just found nice the idea of having the article next to the code in the
whole demo.