Adds dual arm simultaneous planning
It builds a single chaing composed of the left arm and the reversed right
arm, coupled with a revolute joint whose origin is the (safe) contact
point and rotation axis the normal at the expected contact point.

The target frame for the single two-arm chain with that virtual joint in
the middle is the identity since we want to close it into itself.

The IK solution is then sent as the desired position for motion planning
through MoveIt!

cc: @Tabjones, a requirement for this approach is that the processing frame
be left_hand_palm_link instead of left_arm_7_link, since the virtual joint
in between is built using the contact and normal, and it's a lot easier if
they are in the left_hand_palm_link. I don't think there should be any
problem, but if so, let me know, otherwise we will see alive.