Abstract and Sections 1-3 are readable
Those correspond to files intro.tex, related.tex, scope.tex and of course,
the main report.tex. They are quite dense so far, since in Section 3 a lot
of notation and background is already introduced, so that Section 4 can
focus on our extension, contribution, and strategy, with all previous
concepts in mind.

Feel free to edit and make modifications to those files, since I continue
with the remaining sections.

cc: @jeremylwyatt @MarcoGabiccini @Tabjones @memnone

ps: sorry for the delay.